Ensure request log middleware higher than common middleware
Django's common middleware is what ensures requests without a trailing
slash in the URL are redirected (with a 301 status) to URLs with a
trailing slash.  But until this happens, the request is handled as a
404, so it will be logged that way unless the common middleware is run
prior to the log entry being generated.  So this change moves the log
middleware above the common middleware to ensure this happens.