asset scanner: Handle request exceptions when fetching assets
Updates the asset scanner/fetcher to catch any exceptions the
`requests` library raises when attempting to obtain the contents of
various asset files.  The previous behavior was to let those be
raised.

I've changed the `fetch_asset` function, the only place where we
invoke `requests`, to return a string containing the text content of
the response, rather than the response itself.  This is a good change
anyway, since it means the callers of `fetch_asset`, all of which only
wanted the text content anyway, don't have to worry about the details
of the `requests` library.

One thing to note: if an error is caught, then the function returns
the empty string.  This is the simplest thing to do, but it means
we're kind of passing over the error.  If we want to flag it in the
results somehow, then we will have to add some additional code and
make decisions about how exactly we want that to work.